fix(file-upload): refactor fileupload to use params instead of header
Just found the params field today, this change is to go with general SWE practices that a header is a field of an HTTP request or response that passes additional context and metadata about the request or response. For example, a request message can use headers to indicate it's preferred media formats, while a response can use header to indicate the media format of the returned body.